pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bank-agent-llm"
version = "0.1.0"
description = "Local-first AI pipeline to fetch, parse, categorize, and query personal bank statements"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["finance", "llm", "ollama", "etl", "bank-statements", "cli"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    # CLI
    "typer>=0.12",
    "rich>=13.7",
    # Config
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    # Email
    "imapclient>=3.0",
    # PDF / spreadsheet parsing
    "pdfplumber>=0.11",
    "openpyxl>=3.1",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    # HTTP (Ollama API)
    "httpx>=0.27",
    # Resilience (retries for IMAP and Ollama calls)
    "tenacity>=8.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.30",   # mock httpx calls in tests
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
bank-agent = "bank_agent_llm.cli:app"

[project.urls]
Repository = "https://github.com/JuanLara18/bank-agent-llm"
Issues = "https://github.com/JuanLara18/bank-agent-llm/issues"

# ─── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.ruff.format]
quote-style = "double"

# ─── Mypy ────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ─── Pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src"]
omit = ["*/migrations/*", "*/cli.py"]

# ─── Hatch ───────────────────────────────────────────────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["src/bank_agent_llm"]
